requirements.txt
openai==1.95.0
python-dotenv==1.1.1
streamlit==1.46.1s
